[fix] : 답변완료한 상품문의 카운트 안맞는거 해결

+++ src/main/java/shop/itbook/itbookshop/productgroup/productinquiry/dto/response/ProductInquiryCountResponseDto.java
@@ -16,5 +16,5 @@ import lombok.NoArgsConstructor;
 public class ProductInquiryCountResponseDto {
 
     private Long productInquiryCount;
-    private Long repliedProductInquiryCount;
+    private Integer repliedProductInquiryCount;
 }
